requirements.txt
# Core Deep Learning
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0

# Diffusion & Transformers
diffusers>=0.27.0
transformers>=4.38.0
peft>=0.10.0
accelerate>=0.27.0

# Data Processing
Pillow>=10.0.0
datasets>=2.18.0

# Optimization & Training
bitsandbytes>=0.43.0
safetensors>=0.4.0
xformers>=0.0.22

# Configuration & Logging
pyyaml>=6.0
wandb>=0.16.0

# API & CLI
flask>=3.0.0
click>=8.1.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
